# public/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TPLK004 - Sistem Absensi & Manajemen Akademik API",
    "description": "API Documentation untuk Sistem Absensi QR Code dengan Gamifikasi, Manajemen Tugas, dan Analytics. Sistem ini menyediakan endpoint untuk autentikasi, manajemen absensi, tugas, gamifikasi, dan berbagai fitur akademik lainnya.",
    "version": "1.0.0",
    "contact": {
      "name": "TPLK004 Support Team",
      "email": "admin@unpam.ac.id"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter token in format: Bearer {token}"
      },
      "sanctum": {
        "type": "apiKey",
        "in": "cookie",
        "name": "laravel_session",
        "description": "Laravel Sanctum session authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks",
        "description": "Mendapatkan daftar tugas dengan filter",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "course_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["published", "draft", "closed"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create new task (Dosen only)",
        "description": "Membuat tugas baru",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["course_id", "judul", "deskripsi", "jenis", "deadline"],
                "properties": {
                  "course_id": {
                    "type": "integer"
                  },
                  "judul": {
                    "type": "string"
                  },
                  "deskripsi": {
                    "type": "string"
                  },
                  "jenis": {
                    "type": "string",
                    "enum": ["tugas", "quiz", "project", "presentasi", "lainnya"]
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully"
          }
        }
      }
    },
...
  }
}